use RFduinoBLE.deviceName, to set name characteristic
This name characteristic is returned sometimes by iOS rduino library
for the name of the device. This seems to happen only after the device
has been connected to once.
Using the raw advdata approach doesn’t set this name characteristic.
Additionally the Arduino code limits the name to 14 chars so that it can
include at least 1 character of advertisement data.